WAGE: Extract MacWindow
