+++ engines/wage/gui-console.cpp
@@ -51,11 +51,11 @@
 #include "graphics/cursorman.h"
 #include "graphics/fonts/bdf.h"
 #include "graphics/palette.h"
+#include "graphics/macgui/macwindow.h"
 
 #include "wage/wage.h"
 #include "wage/design.h"
 #include "wage/entities.h"
-#include "wage/macwindow.h"
 #include "wage/macmenu.h"
 #include "wage/gui.h"
 #include "wage/world.h"
