+++ engines/wage/macmenu.cpp
@@ -51,8 +51,8 @@
 #include "graphics/primitives.h"
 #include "graphics/font.h"
 #include "graphics/macgui/macwindowmanager.h"
+#include "graphics/macgui/macwindow.h"
 
-#include "wage/macwindow.h"
 #include "wage/macmenu.h"
 
 namespace Wage {
