+++ graphics/macgui/macwindowmanager.cpp
@@ -58,8 +58,8 @@
 #include "graphics/palette.h"
 #include "graphics/primitives.h"
 #include "graphics/macgui/macwindowmanager.h"
+#include "graphics/macgui/macwindow.h"
 
-#include "wage/macwindow.h"
 #include "wage/macmenu.h"
 
 namespace Graphics {
